Add rough initial code execution runtime for JS
Also, experiment with an idea for verifying functionality under
:advanced compilation (test.js). We can hopefully write tests that use
`require` to access the eval fn instead of having to write them in
Javascript, but even if we do use JS it'd be worth it to have that
functionality verified.

The actual thing should have the exported entry point in a separate
namespace that doesn't get packaged with clojars builds.