Add support for automated testing of advanced build capabilities
A bit tricky since the ci-engine.js `goog.provide`s some `cljs.core`
stuff, so sadly we can't run them all together. Basically we do a
vanilla karma build and run that, then build the engine in advanced
mode, and the tests that run against it (which I guess should not
require *anything* themselves) and run those with the ci-engine.js
prepended.