Refactor expose-fn (to export-fn) to simplify generated code
I think the new macro generates simpler JS; the old way created this
lovecraftian nested function call thing. It's possible that Closure
could simplify that structure, but this is still probably better.

Also, the way forward on macros is probably to just re-implement them
and apply them to the input form by hand....