WIP: experiment with a "hybrid" exec/eval approach
Sadly, directly evaluating everything as a script is terribly slow,
going from ~500ms to load the SRD to ~2000ms. Yikes.

Most of the time is spent compiling the forms into javascript, and not
actually on execution, however, so I think we can come up with a hybrid
approach that compiles things like `(let)` forms and directly evaluates
other things. There are definitely big wins to be had with massive
static data like the SRD magic items and spells lists.