First pass with a functional hybrid/eager evaluation approach
This approach is not as fast as the previous mechanism (~1200ms instead
of ~900ms) but all the tests actually pass, so it's at least a *correct*
approach. 600ms shaved off is still a great start, but hopefully there's
another 600ms that we won't have to get by nickles and dimes.