Lazily compile functions
This shaves off 500ms, from ~1200ms to ~700ms! Now we're in the
ballpark!

Most of the small functions compile quickly enough that it shouldn't be
too noticeable when they get lazily compiled during use, and many will
probably not be seen at all—especially those on a class or item, for
example—so I think this is a great tradeoff.