Significantly improve initial load perf with "hybrid" execution + lazy compilation (#3)
* WIP: experiment with a "hybrid" exec/eval approach

Sadly, directly evaluating everything as a script is terribly slow,
going from ~500ms to load the SRD to ~2000ms. Yikes.

Most of the time is spent compiling the forms into javascript, and not
actually on execution, however, so I think we can come up with a hybrid
approach that compiles things like `(let)` forms and directly evaluates
other things. There are definitely big wins to be had with massive
static data like the SRD magic items and spells lists.

* First pass with a functional hybrid/eager evaluation approach

This approach is not as fast as the previous mechanism (~1200ms instead
of ~900ms) but all the tests actually pass, so it's at least a *correct*
approach. 600ms shaved off is still a great start, but hopefully there's
another 600ms that we won't have to get by nickles and dimes.

* Lazily compile functions

This shaves off 500ms, from ~1200ms to ~700ms! Now we're in the
ballpark!

Most of the small functions compile quickly enough that it shouldn't be
too noticeable when they get lazily compiled during use, and many will
probably not be seen at all—especially those on a class or item, for
example—so I think this is a great tradeoff.

* Simplify handling of `false` value returned from a macro

* Ensure scripting-api functions are also properly evaluated

* Remove logging and profiling